// frontend/src/api/models/AgentOut.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AgentOut = {
    avatar: (string | null);
    model_id: number;
    description: (string | null);
    system_prompt: (string | null);
    max_token_size: (number | null);
    human_in_the_loop?: (boolean | null);
    user_id: number;
    id: number;
    created_at: string;
    updated_at: string;
    mcp_ids?: (Array<number> | null);
};