// frontend/src/api/models/AgentMCPOut.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AgentMCPOut = {
    id: number;
    user_id: (number | null);
    mcp_name: string;
    transport: string;
    config_json: (Record<string, any> | null);
    is_builtin: boolean;
    is_enabled: boolean;
    created_at: string;
    updated_at: string;
    has_mcp: boolean;
};